updating the mock api a bit
